backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
openpyxl==3.1.5
xlrd>=2.0.1
google-genai==0.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.4
